+++ mod/readaloud/mod_form.php
@@ -101,19 +101,16 @@ class mod_readaloud_mod_form extends moodleform_mod {
 
     /**
      * This adds completion rules
-     * The values here are just dummies. They don't work in this project until you implement some sort of grading
      * See lib.php readaloud_get_completion_state()
      */
     function add_completion_rules() {
         $mform =& $this->_form;
         $config = get_config(constants::M_COMPONENT);
 
-        //timer options
-        //Add a place to set a mimumum time after which the activity is recorded complete
-        $mform->addElement('static', 'mingradedetails', '', get_string('mingradedetails', constants::M_COMPONENT));
-        $options = array(0 => get_string('none'), 20 => '20%', 30 => '30%', 40 => '40%', 50 => '50%', 60 => '60%', 70 => '70%',
+        //minimum grade
+        $mingradeoptions = array(0 => get_string('none'), 20 => '20%', 30 => '30%', 40 => '40%', 50 => '50%', 60 => '60%', 70 => '70%',
                 80 => '80%', 90 => '90%', 100 => '40%');
-        $mform->addElement('select', 'mingrade', get_string('mingrade', constants::M_COMPONENT), $options);
+        $mform->addElement('select', 'mingrade', get_string('mingrade', constants::M_COMPONENT), $mingradeoptions);
 
         return array('mingrade');
     }